requirements.txt
pytest==7.4.0
pytest-ordering==0.6
selenium==4.8.3
requests==2.28.2
allure-pytest==2.13.2
PyYAML~=6.0
redis==6.4.0